Fix #68 visibility_priority type changed `int` to `long`

+++ app/src/androidTest/java/com/nlab/reminder/macro/InsertDummyInfoToDatabaseMacro.kt
@@ -112,7 +112,7 @@ class InsertDummyInfoToDatabaseMacro {
         scheduleDao.findByComplete(isComplete = false).first().forEach { scheduleDao.delete(it.scheduleEntity) }
 
         inputScheduleEntities.shuffled().forEachIndexed { index, scheduleEntity ->
-            scheduleDao.insert(scheduleEntity.copy(visiblePriority = index))
+            scheduleDao.insert(scheduleEntity.copy(visiblePriority = index.toLong()))
         }
     }
